+++ src/main/java/com/raoulvdberge/refinedstorageaddons/RSAddonsConfig.java
@@ -14,16 +14,6 @@ import java.util.List;
 public final class RSAddonsConfig {
     private Configuration config;
 
-    //region Network Picker
-    public int networkPickerUsage;
-    public boolean networkPickerUsesEnergy;
-    //endRegion
-
-    //region Network Bag
-    public int networkBagUsage;
-    public boolean networkBagUsesEnergy;
-    //endRegion
-
     //region Infinite Wireless Transmitter
     public int infiniteWirelessTransmitterUsage;
     public int infiniteWirelessTransmitterRange;
@@ -38,8 +28,6 @@ public final class RSAddonsConfig {
     //endregion
 
     //region Categories
-    private static final String NETWORK_PICKER = "networkPicker";
-    private static final String NETWORK_BAG = "networkBag";
     private static final String INFINITE_WIRELESS_TRANSMITTER = "infiniteWirelessTransmitter";
     private static final String WIRELESS_CRAFTING_GRID = "wirelessCraftingGrid";
     //endregion
@@ -64,16 +52,6 @@ public final class RSAddonsConfig {
     }
 
     private void loadConfig() {
-        //region Network Picker
-        networkPickerUsesEnergy = config.getBoolean("usesEnergy", NETWORK_PICKER, true, "Whether the Network Picker uses energy");
-        networkPickerUsage = config.getInt("usage", NETWORK_PICKER, 45, 0, Integer.MAX_VALUE, "The energy used by the Network Picker");
-        //endRegion
-
-        //region Network Picker
-        networkBagUsesEnergy = config.getBoolean("usesEnergy", NETWORK_BAG, true, "Whether the Network Bag uses energy");
-        networkBagUsage = config.getInt("usage", NETWORK_BAG, 15, 0, Integer.MAX_VALUE, "The energy used by the Network Bag");
-        //endRegion
-
         //region Infinite Wireless Transmitter
         infiniteWirelessTransmitterUsage = config.getInt("infiniteWirelessTransmitterUsage", INFINITE_WIRELESS_TRANSMITTER, 1000, 0, Integer.MAX_VALUE, "The energy used by the Infinite Wireless Transmitter");
         infiniteWirelessTransmitterRange = config.getInt("infiniteWirelessTransmitterRange", INFINITE_WIRELESS_TRANSMITTER, Integer.MAX_VALUE, 0, Integer.MAX_VALUE, "The range of the Infinite Wireless Transmitter");
@@ -96,10 +74,8 @@ public final class RSAddonsConfig {
     public List<IConfigElement> getConfigElements() {
         List<IConfigElement> list = new ArrayList<>();
 
-        list.addAll(new ConfigElement(config.getCategory(NETWORK_PICKER)).getChildElements());
-        list.addAll(new ConfigElement(config.getCategory(NETWORK_BAG)).getChildElements());
-        list.addAll(new ConfigElement(config.getCategory(INFINITE_WIRELESS_TRANSMITTER)).getChildElements());
-        list.addAll(new ConfigElement(config.getCategory(WIRELESS_CRAFTING_GRID)).getChildElements());
+        list.add(new ConfigElement(config.getCategory(INFINITE_WIRELESS_TRANSMITTER)));
+        list.add(new ConfigElement(config.getCategory(WIRELESS_CRAFTING_GRID)));
 
         return list;
     }